Add auto refresh function.
