+++ src/index.html
@@ -24,12 +24,6 @@
         </button>
       </div>
       <div id="scoreboard" class="scores-border">
-        <p>Name: 100</p>
-        <p>Name: 20</p>
-        <p>Name: 50</p>
-        <p>Name: 78</p>
-        <p>Name: 125</p>
-        <p>Name: 42</p>
       </div>
     </div>
 
@@ -52,7 +46,7 @@
 
         </div>
         <div class="flex justify-end items-center">
-          <button class="button-style" type="button">
+          <button id="submit" class="button-style" type="button">
             <span class="text-xs md:text-sm lg:text-base"> Submit</span>
 
           </button>
